refactor: Rename and optimize search functionality
- Renamed `handleUnifiedSearch` to `handleSearch` for clarity and consistency.
- Improved input schema descriptions, specifying restrictions for search parameters.
- Restructured the retrieved handler to split logic into multiple files
