+++ src/handlers/index.ts
@@ -1,6 +1,6 @@
 export { handleListDatabases } from './listDatabases.js';
 export { handleExecuteQuery } from './executeQuery.js';
-export { handleUnifiedSearch } from './unifiedSearch.js';
+export { handleSearch } from './search.js';
 export { handleExportQuery } from './exportQuery.js';
 export { handleClearCache } from './clearCache.js';
-export { handleRetrieve } from './retrieve.js';
+export { handleRetrieve } from './retrieve/index.js';
